moved all functionality to Query class
