+++ src/Literals/StringLiteral.php
@@ -30,7 +30,7 @@ use WikibaseSolutions\CypherDSL\Types\PropertyTypes\StringType;
  *
  * @see https://neo4j.com/docs/cypher-manual/current/syntax/expressions/#cypher-expressions-string-literals
  */
-class StringLiteral extends Literal implements StringType
+class StringLiteral implements StringType
 {
     use StringTypeTrait;
