+++ src/Query.php
@@ -54,6 +54,14 @@ use WikibaseSolutions\CypherDSL\Types\PropertyTypes\StringType;
 use WikibaseSolutions\CypherDSL\Types\StructuralTypes\NodeType;
 use WikibaseSolutions\CypherDSL\Types\StructuralTypes\PathType;
 use WikibaseSolutions\CypherDSL\Types\StructuralTypes\StructuralType;
+use function get_class;
+use function gettype;
+use function is_bool;
+use function is_float;
+use function is_int;
+use function is_object;
+use function is_string;
+use function method_exists;
 
 /**
  * Builder class for building complex Cypher queries.
@@ -125,25 +133,25 @@ class Query implements QueryConvertable
      * Creates a new literal from the given value. This function automatically constructs the appropriate
      * class based on the type of the value given.
      *
-     * @param integer|float|bool|string $literal The literal to construct
+     * @param mixed $literal The literal to construct
      * @return StringLiteral|Boolean|Decimal
      */
     public static function literal($literal): PropertyType
     {
-        $literalType = gettype($literal);
-
-        switch ($literalType) {
-            case "string":
-                return new StringLiteral($literal);
-            case "boolean":
-                return new Boolean($literal);
-            case "double":
-            case "float":
-            case "integer":
-                return new Decimal($literal);
-            default:
-                throw new InvalidArgumentException("The literal type " . $literalType . " is not supported by Cypher");
+        if (is_string($literal) || (is_object($literal) && method_exists($literal, '__toString'))) {
+            return new StringLiteral((string) $literal);
         }
+
+        if (is_bool($literal)) {
+            return new Boolean($literal);
+        }
+
+        if (is_int($literal) || is_float($literal)) {
+            return new Decimal($literal);
+        }
+
+        $actualType = is_object($literal) ? get_class($literal) : gettype($literal);
+        throw new InvalidArgumentException("The literal type " . $actualType . " is not supported by Cypher");
     }
 
     /**
@@ -155,6 +163,23 @@ class Query implements QueryConvertable
     public static function list(array $expressions): ExpressionList
     {
         return new ExpressionList($expressions);
+    }
+
+    /**
+     * Creates a list of literal expressions.
+     *
+     * @param mixed[] $expressions
+     *
+     * @return ExpressionList
+     */
+    public static function literalList(iterable $literals): ExpressionList
+    {
+        $tbr = [];
+        foreach ($literals as $literal) {
+            $tbr[] = self::literal($literal);
+        }
+
+        return new ExpressionList($tbr);
     }
 
     /**
