+++ tests/Unit/QueryTest.php
@@ -21,7 +21,9 @@
 
 namespace WikibaseSolutions\CypherDSL\Tests\Unit;
 
+use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
+use TypeError;
 use WikibaseSolutions\CypherDSL\Assignment;
 use WikibaseSolutions\CypherDSL\Clauses\Clause;
 use WikibaseSolutions\CypherDSL\Clauses\MatchClause;
@@ -371,6 +373,67 @@ class QueryTest extends TestCase
         $query = new Query();
 
         $this->assertSame("", $query->build());
+    }
+
+    public function testInt(): void
+    {
+        $literal = Query::literal(1);
+        self::assertInstanceOf(Decimal::class, $literal);
+        self::assertEquals('1', $literal->toQuery());
+    }
+
+    public function testFloat(): void
+    {
+        $literal = Query::literal(1.2);
+        self::assertInstanceOf(Decimal::class, $literal);
+        self::assertEquals('1.2', $literal->toQuery());
+    }
+
+    public function testString(): void
+    {
+        $literal = Query::literal('abc');
+        self::assertInstanceOf(StringLiteral::class, $literal);
+        self::assertEquals("'abc'", $literal->toQuery());
+    }
+
+    public function testStringAble(): void
+    {
+        $literal = Query::literal(new class () {
+            public function __toString(): string
+            {
+                return 'stringable abc';
+            }
+        });
+        self::assertInstanceOf(StringLiteral::class, $literal);
+        self::assertEquals("'stringable abc'", $literal->toQuery());
+    }
+
+    public function testBool(): void
+    {
+        $literal = Query::literal(true);
+        self::assertInstanceOf(Boolean::class, $literal);
+        self::assertEquals("true", $literal->toQuery());
+    }
+
+    public function testInvalidLiteral(): void
+    {
+        $literal = Query::literal(true);
+        $this->expectException(InvalidArgumentException::class);
+        Query::literal($literal);
+    }
+
+    public function testFromLiterals(): void
+    {
+        $expressionList = Query::literalList(['a', 'b', 234.3, 1, true, false]);
+
+        $this->assertSame("['a', 'b', 234.3, 1, true, false]", $expressionList->toQuery());
+    }
+
+    public function testFromLiteralsError(): void
+    {
+        $expressionList = Query::literalList(['a', 'b', 234.3, 1, true, false]);
+        $this->expectException(InvalidArgumentException::class);
+        $expressionList = Query::literalList([$expressionList]);
     }
 
     public function testWikiExamples()
